update 1.3.0
finish basic download,due to some reasons,maybe browser's problem,we can't download files at the same time but only one file.
Next update will coming soon,ThreadPool is necessary

+++ src/main/java/priv/henryyu/privatebox/model/response/DataGrid.java
@@ -23,7 +23,7 @@ public class DataGrid<T> implements java.io.Serializable {
 	 * 
 	 */
 
-	private int total = 0;
+	private long total = 0;
 
 	private Collection<T> rows;
 
@@ -35,11 +35,13 @@ public class DataGrid<T> implements java.io.Serializable {
 		this.rows = rows;
 	}
 
-	public int getTotal() {
+	
+
+	public long getTotal() {
 		return total;
 	}
 
-	public void setTotal(int total) {
+	public void setTotal(long total) {
 		this.total = total;
 	}
